+++ lib/plausible/stats/base.ex
@@ -1,6 +1,8 @@
 defmodule Plausible.Stats.Base do
   use Plausible.ClickhouseRepo
   use Plausible
+  use Plausible.Stats.Fragments
+
   alias Plausible.Stats.{Query, Filters}
   alias Plausible.Timezones
   import Ecto.Query
@@ -114,14 +116,12 @@ defmodule Plausible.Stats.Base do
           if value == "(none)" do
             from(
               e in q,
-              where: fragment("not has(?, ?)", field(e, :"meta.key"), ^prop_name)
+              where: not has_key(e, :meta, ^prop_name)
             )
           else
             from(
               e in q,
-              array_join: meta in "meta",
-              as: :meta,
-              where: meta.key == ^prop_name and meta.value == ^value
+              where: has_key(e, :meta, ^prop_name) and get_by_key(e, :meta, ^prop_name) == ^value
             )
           end
 
@@ -129,30 +129,34 @@ defmodule Plausible.Stats.Base do
           if value == "(none)" do
             from(
               e in q,
-              where: fragment("has(?, ?)", field(e, :"meta.key"), ^prop_name)
+              where: has_key(e, :meta, ^prop_name)
             )
           else
             from(
               e in q,
-              left_array_join: meta in "meta",
-              as: :meta,
               where:
-                (meta.key == ^prop_name and meta.value != ^value) or
-                  fragment("not has(?, ?)", field(e, :"meta.key"), ^prop_name)
+                not has_key(e, :meta, ^prop_name) or get_by_key(e, :meta, ^prop_name) != ^value
             )
           end
+
+        {"event:props:" <> prop_name, {:matches, value}} ->
+          regex = page_regex(value)
+
+          from(
+            e in q,
+            where:
+              has_key(e, :meta, ^prop_name) and
+                fragment("match(?, ?)", get_by_key(e, :meta, ^prop_name), ^regex)
+          )
 
         {"event:props:" <> prop_name, {:member, values}} ->
           none_value_included = Enum.member?(values, "(none)")
 
           from(
             e in q,
-            left_array_join: meta in "meta",
-            as: :meta,
             where:
-              (meta.key == ^prop_name and meta.value in ^values) or
-                (^none_value_included and
-                   fragment("not has(?, ?)", field(e, :"meta.key"), ^prop_name))
+              (has_key(e, :meta, ^prop_name) and get_by_key(e, :meta, ^prop_name) in ^values) or
+                (^none_value_included and not has_key(e, :meta, ^prop_name))
           )
 
         {"event:props:" <> prop_name, {:not_member, values}} ->
@@ -160,14 +164,13 @@ defmodule Plausible.Stats.Base do
 
           from(
             e in q,
-            left_array_join: meta in "meta",
-            as: :meta,
             where:
-              (meta.key == ^prop_name and meta.value not in ^values) or
-                (^none_value_included and fragment("has(?, ?)", field(e, :"meta.key"), ^prop_name) and
-                   meta.value not in ^values) or
-                (not (^none_value_included) and
-                   fragment("not has(?, ?)", field(e, :"meta.key"), ^prop_name))
+              (has_key(e, :meta, ^prop_name) and
+                 get_by_key(e, :meta, ^prop_name) not in ^values) or
+                (^none_value_included and
+                   has_key(e, :meta, ^prop_name) and
+                   get_by_key(e, :meta, ^prop_name) not in ^values) or
+                (not (^none_value_included) and not has_key(e, :meta, ^prop_name))
           )
 
         _ ->
@@ -230,34 +233,31 @@ defmodule Plausible.Stats.Base do
   def apply_entry_prop_filter(sessions_q, prop_name, {:is, "(none)"}) do
     from(
       s in sessions_q,
-      where: fragment("not has(?, ?)", field(s, :"entry_meta.key"), ^prop_name)
+      where: not has_key(s, :entry_meta, ^prop_name)
     )
   end
 
   def apply_entry_prop_filter(sessions_q, prop_name, {:is, value}) do
     from(
       s in sessions_q,
-      array_join: meta in "entry_meta",
-      as: :meta,
-      where: meta.key == ^prop_name and meta.value == ^value
+      where:
+        has_key(s, :entry_meta, ^prop_name) and get_by_key(s, :entry_meta, ^prop_name) == ^value
     )
   end
 
   def apply_entry_prop_filter(sessions_q, prop_name, {:is_not, "(none)"}) do
     from(
       s in sessions_q,
-      where: fragment("has(?, ?)", field(s, :"entry_meta.key"), ^prop_name)
+      where: has_key(s, :entry_meta, ^prop_name)
     )
   end
 
   def apply_entry_prop_filter(sessions_q, prop_name, {:is_not, value}) do
     from(
       s in sessions_q,
-      left_array_join: meta in "entry_meta",
-      as: :meta,
       where:
-        (meta.key == ^prop_name and meta.value != ^value) or
-          fragment("not has(?, ?)", field(s, :"entry_meta.key"), ^prop_name)
+        not has_key(s, :entry_meta, ^prop_name) or
+          get_by_key(s, :entry_meta, ^prop_name) != ^value
     )
   end
 
@@ -484,28 +484,28 @@ defmodule Plausible.Stats.Base do
     end
   end
 
-  def utc_boundaries(%Query{period: "realtime"}, site) do
+  def utc_boundaries(%Query{period: "realtime", now: now}, site) do
     last_datetime =
-      NaiveDateTime.utc_now()
+      now
       |> Timex.shift(seconds: 5)
       |> beginning_of_time(site.native_stats_start_at)
       |> NaiveDateTime.truncate(:second)
 
     first_datetime =
-      NaiveDateTime.utc_now() |> Timex.shift(minutes: -5) |> NaiveDateTime.truncate(:second)
+      now |> Timex.shift(minutes: -5) |> NaiveDateTime.truncate(:second)
 
     {first_datetime, last_datetime}
   end
 
-  def utc_boundaries(%Query{period: "30m"}, site) do
+  def utc_boundaries(%Query{period: "30m", now: now}, site) do
     last_datetime =
-      NaiveDateTime.utc_now()
+      now
       |> Timex.shift(seconds: 5)
       |> beginning_of_time(site.native_stats_start_at)
       |> NaiveDateTime.truncate(:second)
 
     first_datetime =
-      NaiveDateTime.utc_now() |> Timex.shift(minutes: -30) |> NaiveDateTime.truncate(:second)
+      now |> Timex.shift(minutes: -30) |> NaiveDateTime.truncate(:second)
 
     {first_datetime, last_datetime}
   end
@@ -525,16 +525,15 @@ defmodule Plausible.Stats.Base do
     {first_datetime, last_datetime}
   end
 
-  @replaces %{
-    ~r/\*\*/ => ".*",
-    ~r/(?<!\.)\*/ => "[^/]*",
-    "(" => "\\(",
-    ")" => "\\)"
-  }
   def page_regex(expr) do
-    Enum.reduce(@replaces, "^#{expr}$", fn {pattern, replacement}, regex ->
-      String.replace(regex, pattern, replacement)
-    end)
+    escaped =
+      expr
+      |> Regex.escape()
+      |> String.replace("\\|", "|")
+      |> String.replace("\\*\\*", ".*")
+      |> String.replace("\\*", ".*")
+
+    "^#{escaped}$"
   end
 
   defp split_goals(clauses, map_fn \\ &Function.identity/1) do