+++ test/plausible/imported/buffer_test.exs
@@ -1,14 +1,14 @@
-defmodule Plausible.Google.BufferTest do
+defmodule Plausible.Imported.BufferTest do
   use Plausible.DataCase, async: false
 
   import Ecto.Query
-  alias Plausible.Google.Buffer
+  alias Plausible.Imported.Buffer
 
   setup [:create_user, :create_new_site, :set_buffer_size]
 
   defp set_buffer_size(_setup_args) do
-    google_setting = Application.get_env(:plausible, :google)
-    patch_env(:google, Keyword.put(google_setting, :max_buffer_size, 10))
+    imported_setting = Application.fetch_env!(:plausible, :imported)
+    patch_env(:imported, Keyword.put(imported_setting, :max_buffer_size, 10))
     :ok
   end